Implement full .hbs -> .gjs codemod
Hardcoded a bunch of thins to make it work with Discourse – the
chat plugin to be specific.

We have a bunch of custom resolver rules that needs to be ported
over to make this work.

The overall strategry should be generalizable. We have a bunch of
custom resolver logic that needs to be ported over, the average app
can probably try to share code with the Embroider resolver – or use
the Resolver from `@embroider/core` directly with `.resolver.json`.

See https://github.com/discourse/discourse/pull/24260 for how this
code was used in context.